+++ cereal/log.capnp
@@ -2413,6 +2413,7 @@ struct Event {
     uiDebug @102 :UIDebug;
 
     # *********** debug ***********
+    testJoystick @52 :Joystick;
     roadEncodeData @86 :EncodeData;
     driverEncodeData @87 :EncodeData;
     wideRoadEncodeData @88 :EncodeData;
@@ -2482,6 +2483,5 @@ struct Event {
     uiPlanDEPRECATED @106 :UiPlan;
     liveLocationKalmanDEPRECATED @72 :LiveLocationKalman;
     liveTracksDEPRECATED @16 :List(LiveTracksDEPRECATED);
-    testJoystickDEPRECATED @52 :Joystick;
   }
 }
