+++ tools/joystick/README.md
@@ -2,33 +2,33 @@
 
 **Hardware needed**: device running openpilot, laptop, joystick (optional)
 
-With joystickd, you can connect your laptop to your comma device over the network and debug controls using a joystick or keyboard.
-joystickd uses [inputs](https://pypi.org/project/inputs) which supports many common gamepads and joysticks.
+With joystick_control, you can connect your laptop to your comma device over the network and debug controls using a joystick or keyboard.
+joystick_control uses [inputs](https://pypi.org/project/inputs) which supports many common gamepads and joysticks.
 
 ## Usage
 
-The car must be off, and openpilot must be offroad before starting `joystickd`.
+The car must be off, and openpilot must be offroad before starting `joystick_control`.
 
 ### Using a keyboard
 
-SSH into your comma device and start joystickd with the following command:
+SSH into your comma device and start joystick_control with the following command:
 
 ```shell
-tools/joystick/joystickd.py --keyboard
+tools/joystick/joystick_control.py --keyboard
 ```
 
 The available buttons and axes will print showing their key mappings. In general, the WASD keys control gas and brakes and steering torque in 5% increments.
 
 ### Joystick on your comma three
 
-Plug the joystick into your comma three aux USB-C port. Then, SSH into the device and start `joystickd.py`.
+Plug the joystick into your comma three aux USB-C port. Then, SSH into the device and start `joystick_control.py`.
 
 ### Joystick on your laptop
 
-In order to use a joystick over the network, we need to run joystickd locally from your laptop and have it send `testJoystick` packets over the network to the comma device.
+In order to use a joystick over the network, we need to run joystick_control locally from your laptop and have it send `testJoystick` packets over the network to the comma device.
 
 1. Connect a joystick to your PC.
-2. Connect your laptop to your comma device's hotspot and open a new SSH shell. Since joystickd is being run on your laptop, we need to write a parameter to let controlsd know to start in joystick debug mode:
+2. Connect your laptop to your comma device's hotspot and open a new SSH shell. Since joystick_control is being run on your laptop, we need to write a parameter to let controlsd know to start in joystick debug mode:
    ```shell
    # on your comma device
    echo -n "1" > /data/params/d/JoystickDebugMode
@@ -38,11 +38,11 @@ In order to use a joystick over the network, we need to run joystickd locally fr
    # on your comma device
    cereal/messaging/bridge {LAPTOP_IP} testJoystick
    ```
-4. Start joystickd on your laptop in ZMQ mode.
+4. Start joystick_control on your laptop in ZMQ mode.
    ```shell
    # on your laptop
    export ZMQ=1
-   tools/joystick/joystickd.py
+   tools/joystick/joystick_control.py
    ```
 
 ---
